Remove sql_clients/sql_utils module
This junk drawer module has been cleaned out, and the proper home
for the remaining helper function is perfectly obvious.
